feat: add axios in products component
